Prepare the repository for open-source release
Licensing and attribution
- Relicense to MIT only. Upstream facebook/akd offers MIT OR Apache-2.0, so
  this exercises the MIT option; Meta's copyright notice is preserved in every
  derived file, as MIT requires. LICENSE-MIT -> LICENSE, LICENSE-APACHE removed,
  `license = "MIT"` across all four manifests.
- Add a license header to the 77 source files that had none. All 168 tracked
  .rs files now carry one: 159 MIT, 9 retaining the Espresso Systems
  (HyperPlonk) header on vendored arithmetic.
- Add NOTICE describing what is derived from facebook/akd and from
  EspressoSystems/hyperplonk, and what changed.
- Add SECURITY.md: unaudited crypto, non-ceremonial trusted setup, fixed VRF
  seed fallback, optional-and-off transport security.
- Point crate metadata at this repository instead of facebook/akd; drop Meta's
  Codecov badge from xtask/README.md.
- Rewrite CONTRIBUTING.md, which directed contributors to Facebook's CLA and
  bug bounty, and correct the Code of Conduct's enforcement address, which was
  a Meta mailbox that does not reach this fork's maintainers.
- Add CITATION.cff and a README citation section for ePrint 2026/1681.

Build and reproducibility
- Track Cargo.lock, pin the toolchain in rust-toolchain.toml, declare MSRV
  1.88 (nova-snark's floor), and pin the divan fork to the revision the
  published benchmarks actually used.
- Document the undocumented system prerequisites (protoc, libclang) including
  the macOS DYLD_FALLBACK_LIBRARY_PATH workaround.

Fixes found during the cleanup
- distributed_srs.rs gated its rayon paths on `feature = "parallel"`, which
  akd does not define, so every parallel branch was dead and SRS generation
  silently ran serial. NOTE: this may move the published setup timing.
- akd_core could not build standalone: `proto` was ungated while the protobuf
  dependency is optional, and `aegon_crypto` -- the entire KZH-k
  implementation -- was gated behind `feature = "protobuf"`, so the crate's
  core cryptography vanished under default features and its own test run
  exercised none of it. Gate each on the right feature and add the missing
  `std` feature the SRS file I/O needs.
- gen_srs_for_testing did a non-atomic check-then-write on a shared cache
  file, so concurrent tests deserialized partial files. Write to a temp file
  and rename.
- Remove five leftover `[EPOCH-INSTR]` eprintln! calls that logged on every
  publish.

Tests, lints, CI
- `cargo test -p akd` and `cargo test -p akd_core` are green. The 74 upstream
  SEEMless/Merkle tests move behind an off-by-default `upstream_tests`
  feature; they exercise a backend Aegon replaced. Eight genuine failures are
  marked #[ignore] with the reason inline and tabled in the README.
- Mark the five crate-level doctests that call the legacy AKD entry points
  `ignore`, and add a doc warning pointing at aegon_facade, since those
  entry points are unimplemented on this backend.
- Run rustfmt across the workspace (it had never been run on the Aegon code).
- Record the clippy backlog as an explicit allow-list in [workspace.lints] so
  clippy is clean and enforced against new problems.
- Add .github/workflows/ci.yml: fmt, clippy, build and test on Linux, both
  feature states.
- Untrack 17 MB of crashed-run cluster logs.

+++ akd_core/src/aegon_crypto/pcs/kzhk/srs.rs
@@ -1,3 +1,8 @@
+// Copyright (c) The Aegon Authors.
+//
+// This source code is licensed under the MIT license found in the
+// LICENSE file in the root directory of this source tree.
+
 //! KZH-k structured reference string (SRS).
 //!
 //! Implements the setup described in Figure 14 of ePrint 2025/1580: for
@@ -485,7 +490,7 @@ pub fn ceil_k_root_scaled(n: u128, k: u32) -> u128 {
     // Search m in [k*r_floor, k*(r_floor+1)] s.t. m is the smallest with (m/k)^k >=
     // n. Equivalently: m^k >= n * k^k.
     let lo = (r_floor).saturating_mul(k as u128);
-    let hi = ((r_floor + 1) as u128).saturating_mul(k as u128);
+    let hi = (r_floor + 1).saturating_mul(k as u128);
 
     let target = BigUint::from(n) * pow_big(&BigUint::from(k as u128), k);
     let mut l = BigUint::from(lo);
